--- a/src/methodical/impl.clj
+++ b/src/methodical/impl.clj
@@ -17,16 +17,15 @@
             [methodical.impl.multifn
              [cached :as multifn.cached]
              [standard :as multifn.standard]]
-            [methodical.impl.standard :as impl.standard])
-  (:import methodical.impl.standard.StandardMultiFn
-           [methodical.interface Cache Dispatcher MethodCombination MethodTable MultiFnImpl]))
+            [methodical.impl.standard :as impl.standard]
+            [methodical.interface :as i]))
 
 ;;;; ### Method Combinations
 
 (defn clojure-method-combination
   "Simple method combination strategy that mimics the way vanilla Clojure multimethods combine methods; that is, to say,
   not at all. Like vanilla Clojure multimethods, this method combination only supports primary methods."
-  ^MethodCombination []
+  []
   (combo.clojure/->ClojureMethodCombination))
 
 (defn clos-method-combination
@@ -34,19 +33,19 @@
   Supports `:before`, `:after`, and `:around` auxiliary methods. The values returned by `:before` and `:after` methods
   are ignored. Primary methods and around methods get an implicit `next-method` arg (see Methodical dox for more on
   what this means)."
-  ^MethodCombination []
+  []
   (combo.clos/->CLOSStandardMethodCombination))
 
 (defn thread-first-method-combination
   "Similar the the standard CLOS-style method combination, but threads the result of each `:before` and `:after`
   auxiliary methods, as well as the primary method, as the *first* arg of subsequent method invocations."
-  ^MethodCombination []
+  []
   (combo.threaded/threading-method-combination :thread-first))
 
 (defn thread-last-method-combination
   "Similar the the standard CLOS-style method combination, but threads the result of each `:before` and `:after`
   auxiliary methods, as well as the primary method, as the *last* arg of subsequent method invocations."
-  ^MethodCombination []
+  []
   (combo.threaded/threading-method-combination :thread-last))
 
 
@@ -70,26 +69,26 @@
   "Based on the CLOS `progn` method combination. Sequentially executes *all* applicable primary methods, presumably for
   side-effects, in order from most-specific to least-specific; returns the value returned by the least-specific
   method. `do` method combinations support `:around` auxiliary methods, but not `:before` or `:after` methods."
-  ^MethodCombination []
+  []
   (combo.operator/operator-method-combination :do))
 
 (defn min-method-combination
   "Based on the CLOS method combination of the same name. Executes *all* applicable primary methods, returning the
   minimum value returned by any implementation. Like `do` method combinations, `min` supports `:around` auxiliary
   methods, but not `:before` or `:after`."
-  ^MethodCombination []
+  []
   (combo.operator/operator-method-combination :min))
 
 (defn max-method-combination
   "Executes *all* applicable primary methods, and returns the maximum value returned by any one implemenation. Same
   constraints as othe CLOS operator-style method combinations."
-  ^MethodCombination []
+  []
   (combo.operator/operator-method-combination :max))
 
 (defn +-method-combination
   "Executes *all* applicable primary methods, returnings the sum of the values returned by each method. Same constraints
   as othe CLOS operator-style method combinations."
-  ^MethodCombination []
+  []
   (combo.operator/operator-method-combination :+))
 
 (defn seq-method-combination
@@ -97,14 +96,14 @@
   the method invocations. Inspired by CLOS `nconc` and `append` method combinations, but unlike those, this
   combination returns a completely lazy sequence. Like other CLOS-operator-inspired method combinations, this
   combination currently supports `:around` methods, but not `:before` or `:after` methods."
-  ^MethodCombination []
+  []
   (combo.operator/operator-method-combination :seq))
 
 (defn concat-method-combination
   "Like the `seq-method-combination`, but concatenates all the results together.
 
     seq-method-combination : map :: concat-method-combination : mapcat"
-  ^MethodCombination []
+  []
   (combo.operator/operator-method-combination :concat))
 
 
@@ -112,13 +111,13 @@
   "Invoke *all* applicable primary methods, from most-specific to least-specific; reducing the results as if by `and`.
   Like `and`, this method invocation short-circuits if any implementation returns a falsey value. Otherwise, this
   method returns the value returned by the last method invoked."
-  ^MethodCombination []
+  []
   (combo.operator/operator-method-combination :and))
 
 (defn or-method-combination
   "Like the `and` combination, but combines result as if by `or`; short-circuits after the first matching primary method
   returns a truthy value."
-  ^MethodCombination []
+  []
   (combo.operator/operator-method-combination :or))
 
 
@@ -129,10 +128,10 @@
   multimethods handle multimethod dispatch, with support for a custom `hierarchy`, `default-value` and map of
   `prefers`."
   {:style/indent 1}
-  ^Dispatcher [dispatch-fn & {:keys [hierarchy default-value prefers]
-                              :or   {hierarchy     #'clojure.core/global-hierarchy
-                                     default-value :default
-                                     prefers       {}}}]
+  [dispatch-fn & {:keys [hierarchy default-value prefers]
+                  :or   {hierarchy     #'clojure.core/global-hierarchy
+                         default-value :default
+                         prefers       {}}}]
   {:pre [(ifn? dispatch-fn) (var? hierarchy) (map? prefers)]}
   (dispatcher.standard/->StandardDispatcher dispatch-fn hierarchy default-value prefers))
 
@@ -140,9 +139,9 @@
   "A Dispatcher that always considers *all* primary and auxiliary methods to be matches; does not calculate dispatch
   values for arguments when invoking. Dispatch values are still used to sort methods from most- to least- specific,
   using `hierarchy` and map of `prefers`."
-  ^Dispatcher [& {:keys [hierarchy prefers]
-                  :or   {hierarchy #'clojure.core/global-hierarchy
-                         prefers   {}}}]
+  [& {:keys [hierarchy prefers]
+      :or   {hierarchy #'clojure.core/global-hierarchy
+             prefers   {}}}]
   (dispatcher.everything/->EverythingDispatcher hierarchy prefers))
 
 
@@ -150,20 +149,20 @@
 
 (defn clojure-method-table
   "Create a new Clojure-style method table. Clojure-style method tables only support primary methods."
-  (^MethodTable  []
+  ([]
    (clojure-method-table {}))
 
-  (^MethodTable [m]
+  ([m]
    {:pre [(map? m)]}
    (method-table.clojure/->ClojureMethodTable m)))
 
 
 (defn standard-method-table
   "Create a new standard method table that supports both primary and auxiliary methods."
-  (^MethodTable []
+  ([]
    (standard-method-table {} {}))
 
-  (^MethodTable [primary aux]
+  ([primary aux]
    {:pre [(map? primary) (map? aux)]}
    (method-table.standard/->StandardMethodTable primary aux)))
 
@@ -172,17 +171,17 @@
 
 (defn simple-cache
   "Create a basic dumb cache. The simple cache stores"
-  (^Cache []
+  ([]
    (simple-cache {}))
 
-  (^Cache [m]
+  ([m]
    (cache.simple/->SimpleCache (atom m))))
 
 (defn watching-cache
   "Wrap `cache` in a `WatchingCache`, which clears the cache whenever one of the watched `references` (such as vars or
   atoms) changes. Intended primarily for use with 'permanent' MultiFns, such as those created with `defmulti`; this is
   rarely needed or wanted for transient multifns."
-  ^Cache [cache references]
+  [cache references]
   (cache.watching/add-watches cache references))
 
 
@@ -190,16 +189,16 @@
 
 (defn standard-multifn-impl
   "Create a basic multifn impl using method combination `combo`, dispatcher `dispatcher`, and `method-table`."
-  ^MultiFnImpl [combo dispatcher method-table]
-  {:pre [(instance? MethodCombination combo)
-         (instance? Dispatcher dispatcher)
-         (instance? MethodTable method-table)]}
+  [combo dispatcher method-table]
+  {:pre [(satisfies? i/MethodCombination combo)
+         (satisfies? i/Dispatcher dispatcher)
+         (satisfies? i/MethodTable method-table)]}
   (multifn.standard/->StandardMultiFnImpl combo dispatcher method-table))
 
 (defn default-multifn-impl
   "Create a basic multifn impl using default choices for method combination, dispatcher, and method table."
   {:arglists '([dispatch-fn & {:keys [hierarchy default-value prefers]}])}
-  ^MultiFnImpl [dispatch-fn & dispatcher-options]
+  [dispatch-fn & dispatcher-options]
   (standard-multifn-impl
    (thread-last-method-combination)
    (apply standard-dispatcher dispatch-fn dispatcher-options)
@@ -208,7 +207,7 @@
 (defn clojure-multifn-impl
   "Create a mulitfn impl that largely behaves the same way as a vanilla Clojure multimethod."
   {:arglists '([dispatch-fn & {:keys [hierarchy default-value prefers method-table]}])}
-  ^MultiFnImpl [dispatch-fn & {:keys [method-table], :or {method-table {}}, :as options}]
+  [dispatch-fn & {:keys [method-table], :or {method-table {}}, :as options}]
   (let [dispatcher-options (apply concat (select-keys options [:hierarchy :default-value :prefers]))]
     (standard-multifn-impl
      (clojure-method-combination)
@@ -221,9 +220,9 @@
   `:before` and `:after` methods are ignored, rather than threaded. Primary and `:around` methods each get an implicit
   `next-method` arg."
   {:arglists '([dispatch-fn & {:keys [hierarchy default-value prefers primary-method-table aux-method-table]}])}
-  ^MultiFnImpl [dispatch-fn & {:keys [primary-method-table aux-method-table],
-                                     :or   {primary-method-table {}, aux-method-table {}}
-                                     :as   options}]
+  [dispatch-fn & {:keys [primary-method-table aux-method-table],
+                  :or   {primary-method-table {}, aux-method-table {}}
+                  :as   options}]
   (let [dispatcher-options (apply concat (select-keys options [:hierarchy :default-value :prefers]))]
     (standard-multifn-impl
      (clos-method-combination)
@@ -233,10 +232,10 @@
 (defn cached-multifn-impl
   "Wrap a `MultiFnImpl` in a `CachedMultiFnImpl`, which adds caching to calculated effective methods. The cache itself
   is swappable with other caches that implement different strategies."
-  (^MultiFnImpl [impl]
+  ([impl]
    (cached-multifn-impl impl (simple-cache)))
 
-  (^MultiFnImpl [impl cache]
+  ([impl cache]
    (multifn.cached/->CachedMultiFnImpl impl cache)))
 
 
@@ -245,21 +244,21 @@
 (defn uncached-multifn
   "Create a new Methodical multifn using `impl` as the multifn implementation; `impl` itself should implement
   `MultiFnImpl`. DOES NOT CACHE EFFECTIVE METHODS -- use `multifn` instead, unless you like slow dispatch times."
-  (^StandardMultiFn [impl]
+  ([impl]
    (uncached-multifn impl nil))
 
-  (^StandardMultiFn [impl mta]
+  ([impl mta]
    (impl.standard/->StandardMultiFn impl mta)))
 
 (defn multifn
   "Create a new *cached* Methodical multifn using `impl` as the multifn implementation."
-  (^StandardMultiFn [impl]
+  ([impl]
    (multifn impl nil))
 
-  (^StandardMultiFn [impl mta]
+  ([impl mta]
    (multifn impl mta (simple-cache)))
 
-  (^StandardMultiFn [impl mta cache]
+  ([impl mta cache]
    (uncached-multifn (cached-multifn-impl impl cache) mta)))
 
 (def ^{:arglists (:arglists (meta #'default-multifn-impl))}