… src/methodical/after.txt
… src/methodical/before.txt
… src/methodical/interface.clj
… test/methodical/impl/cache/watching_test.clj
… test/methodical/impl/dispatcher/standard_test.clj