Benchmark for #20. 

+ `definterface` => `defprotocol`
+ `deftype+` => `deftype`
+ remove type hints
+ direct calls => protocol calls

Benchmark
+ simple: 238,312268 ns vs 4,945147 µs   **x10**
+ big-hierarchy: 19,027996 ns vs 2,401364 µs **x100**
+ big-hierarchy +clojure: 16,850978 ns vs 1,476671 µs  **x100**

May be it is possible to preserve type hints in clojurescript and save some performance.

Is it reasonable price for clojurescript?