@camsaul 2019-09-11
Do you mean to say performance is 10-100x slower with those changes?

I was thinking for CLJS support that using `#?` to use either `definterface/defprotocol` and `deftype/defrecord` (etc.) would be the best way to go about doing it. That way Clojure code could keep the fast performance. Not sure how easy that would be