@darkleaf 2019-09-12
> Do you mean to say performance is 10-100x slower with those changes?

yes

> using #?

It's so ugly for type hints:
```clojure
;; before
([^MultiFnImpl impl]
 ((effective-method impl (.dispatch-value ^Dispatcher (.dispatcher impl)))))
;;after. slow cljc
([impl]
 ((effective-method impl (i/dispatch-value (i/dispatcher impl)))))
;;fast cljc
([#?(:clj ^MultiFnImpl) impl]
 ((effective-method impl (#?@(:clj [.dispatch-value ^Dispatcher] :cljs [i/dispatch-value]) (#?(:clj .dispatcher :cljs i/dispatcher) impl)))))
```

Or we can use a bit less ugly but wery strange approach:

```clojure
#?(:clj
   (definterface+ Foo
     (foo [x]))
   :cljs
   (def Foo))

(deftype Bar []
  #?(:clj Foo
     :cljs Object)
  (foo [_] :ok))

(let [^Foo b (Bar.)]
  (.foo b))
```

So this kind of optimizations is so tricky.